requirements.txt
Flask==2.0.3
python-dotenv==0.19.2
requests==2.27.1
elasticsearch==7.16.0
PyJWT==2.3.0
Flask-Limiter==2.3.0
Flask-Talisman==1.0.0
Faker==13.3.4
opentracing==2.4.0
jaeger-client==4.3.0
rstr==2.2.6
Flask-SQLAlchemy==2.5.1
Flask-Bcrypt==1.0.1
Werkzeug==2.0.3
Flask-CORS
numpy==2.2.3
scikit-learn==1.6.1
SQLAlchemy==1.4.49
flask_jwt_extended
gunicorn

